backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
starlette==0.41.3

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis
redis==5.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# Validation & settings
pydantic==2.10.3
pydantic-settings==2.6.1

# HTTP client
httpx==0.28.1

# Task queue
celery==5.4.0

# Encryption (LinkedIn token)
cryptography==44.0.0

# Storage
boto3==1.35.80

# Email
sendgrid==6.11.0

# Rate limiting
slowapi==0.1.9

# Observability
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
